+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased](https://github.com/pusher/chatkit-server-go/compare/2.1.0...HEAD)
 
+### Fixes
+
+- Make it possible to clear a previously defined push notification title override
+
 ## [2.1.0](https://github.com/pusher/chatkit-server-go/compare/2.1.0...2.0.0)
 
 ### Additions